[yaml] refactoring: use dependency on 'intellij.yaml.backend' module instead of v2 dependency on YAML plugin (IJPL-176102)
Technically, currently this is the same, because 'intellij.yaml.backend' module is registered as the embedded module of YAML plugin, so it's loaded by the main plugin classloader. However, dependency on the module is more explicit. Also, we may need to temporarily convert 'intellij.yaml.backend' to an optional module to reuse the same plugin.xml for the frontend process.

GitOrigin-RevId: bf96ed0c60cb335fdb0e592a619e1425cb322f2b